Fix region crossings and access to inventory after changing regions within the same simulator

+++ OpenSim/Region/Environment/Scenes/SceneCommunicationService.cs
@@ -653,9 +653,10 @@ namespace OpenSim.Region.Environment.Scenes
                             SendCloseChildAgentConnections(avatar.UUID,avatar.GetKnownRegionList());
                         }
                         // if (teleport success) // seems to be always success here
-                        // the user may change thier profile information in other region,
+                        // the user may change their profile information in other region,
                         // so the userinfo in UserProfileCache is not reliable any more, delete it
-                        m_commsProvider.UserProfileCacheService.RemoveUser(avatar.UUID);
+                        if (avatar.Scene.NeedSceneCacheClear(avatar.UUID))
+                            m_commsProvider.UserProfileCacheService.RemoveUser(avatar.UUID);
                         m_log.InfoFormat("User {0} is going to another region, profile cache removed", avatar.UUID);
                     }
                     else
